Turn on -Wall (with "-Wno-sign-compare" for now) on src/.

+++ src/Encodings/codepage.cpp
@@ -64,7 +64,7 @@ void Codepage::JisEncodeString(const char* src, char* buf,
 
 void Codepage::JisDecodeString(const char* src, char* buf,
                                size_t buflen) const {
-  int srclen = std::strlen(src), i = 0, j = 0;
+  size_t srclen = std::strlen(src), i = 0, j = 0;
   while (i < srclen && j < buflen) {
     unsigned int c1 = (unsigned char) src[i++];
     if ((c1 >= 0x81 && c1 < 0xa0) || (c1 >= 0xe0 && c1 < 0xf0))
